+++ src/prelude.rs
@@ -1,18 +1,6 @@
 pub use crate::{
-    assets::*,
-    audio::*,
-    camera::*,
-    config::*,
-    debug::*,
-    input::*,
-    loading::*,
-    localization::*,
-    metadata::*,
-    platform::*,
-    session::*,
-    ui::{input::MenuAction, *},
-    utils::*,
-    *,
+    assets::*, audio::*, camera::*, config::*, debug::*, input::*, loading::*, localization::*,
+    metadata::*, platform::*, session::*, ui::*, utils::*, *,
 };
 pub use jumpy_core::bevy_prelude::*;
